[general] updated TODO.txt misc casting fixes. updated docs dates...
[toolbar]
added doublebuffer, vertical, nopartial, mixedbuttons, multicolumn, check, group, dropdown, autosize, noprefix, showtext, wholedropdown styles.

[button]
added split, commandlink styles (only testing atm, these may be removed & put in a new combined button control)
added /xdid -n command to support new styles.
added hotchange, dropdown callback events.

+++ Classes/dcxcontrol.cpp
@@ -252,7 +252,7 @@ std::tuple<NoTheme, WindowStyle, WindowExStyle> DcxControl::parseGeneralControlS
 			ExStyles |= WS_EX_TRANSPARENT;
 			break;
 		case L"hidden"_hash:
-			Styles &= static_cast<DWORD>(~WS_VISIBLE);
+			Styles &= ~WS_VISIBLE;
 			break;
 		case L"alpha"_hash:
 			m_bAlphaBlend = true;
